backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9dj-database-url>=2.1.0